Chore: Split move_execution.rs into multiple files
